[level unrated] Title: 옷가게 할인 받기, Time: 0.01 ms, Memory: 4.2 MB -BaekjoonHub

+++ 프로그래머스/unrated/120818. 옷가게 할인 받기/옷가게 할인 받기.cpp
@@ -15,5 +15,5 @@ int solution(int price) {
          discount = 0.2;
     }
     double answer = price - (price * discount);
-    return (int)answer;
+    return answer;
 }
